Backend/requirements.txt
# pip install -r requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-migrate
Flask-JWT-Extended
Werkzeug
python-dotenv
PyJWT
Pillow